Ingest JSON texts
The schema has been cleaned up so that text_nodes now belong
exclusively to the leaf-level part of the work fragment. In practice,
this means that they belong to exemplars, but this change means that we
can collate text_nodes as needed regardless of where we are in the work
hierarchy.