Start working on moving all Exemplar functionality to Versions
From the CTS definition of exemplar:

> The HMT has been using Exemplars to represent specific derivations
of a CTS version created for the purpose of analysis and unambiguous
citation at the level of the token. E.g. a tokenization by morphemes.
This might be more of a derivation than an interpretation but it has
elements of both.

http://capitains.org/pages/vocabulary

We don't really need exemplars yet, and they just add a layer of
interference that is starting to make the whole project more
difficult to deal with.