Start using DataSchema to process CTS XML
We won't be able to use this with the full documents the way that we can for the __cts__.xml metadata files, but we should be able to query the documents' teiHeader elements, for example, and then use the DataSchemata that we declare to get refsDecl information etc.

Where this will get really powerful is when we then use the refsDecl information to query an xml_document for text_nodes, and then process the text_nodes via a DataSchema.

This should, I think, make it much easier to decorate nodes with speaker information etc.