Improve handling of TEI XML versions
Pausanias (tlg0525) is the proof of concept here, and there are
a few things that will need to be updated when we expand beyond prose texts like the Periegesis.

But for now, these changes dramatically improve how we handle these texts.

For example, the infrastructure is now in place to pull out text nodes from a version wholesale, without needing to do much in the way of modification or cleaning. This allows us to preserve markup that has been getting lost over the past few years in other versions of OC