Set up named entities database tables
Entity references have a foreign key constraint on entities, but the references and entities themselves are not constrained by version, exemplar, text node, etc.

The thinking behind this is that entities will most often be looked up by URN, so we'd just be creating extra hassle and difficulty by imposing additional constraints.